+++ services/frontend/src/app/layout/digital-services-footprint/digital-services-footprint-header/link-create-popup/link-create-popup.component.ts
@@ -1,4 +1,12 @@
-import { Component, EventEmitter, inject, input, Input, Output } from "@angular/core";
+import {
+    Component,
+    computed,
+    EventEmitter,
+    inject,
+    input,
+    Input,
+    Output,
+} from "@angular/core";
 import { ClipboardService } from "ngx-clipboard";
 
 @Component({
@@ -12,8 +20,19 @@ export class LinkCreatePopupComponent {
     sharedLink = input<string>("");
     expiryDate = input<Date | null>(null);
     @Output() outClose = new EventEmitter<void>();
+    @Output() outExtendDate = new EventEmitter<void>();
     isLinkCopied = false;
-    loading = false;
+    extendLinkDisabled = computed(() => {
+        if (this.expiryDate()) {
+            const todayUtc = new Date();
+            todayUtc.setUTCHours(23, 59, 0, 0);
+            const diffInDays =
+                (this.expiryDate()!.getTime() - todayUtc.getTime()) / 86400000; // 1000*60*60*24
+            return diffInDays > 59;
+        } else {
+            return true;
+        }
+    });
 
     closePopup(): void {
         this.outClose.emit();
@@ -29,5 +48,9 @@ export class LinkCreatePopupComponent {
         this.isLinkCopied = true;
         const url = this.sharedLink();
         this.clipboardService.copy(url);
+    }
+
+    extendDate(): void {
+        this.outExtendDate.emit();
     }
 }